Move s3 and web directory to under the content directory.
